assignments/finpro/requirements.txt
torch
mujoco==3.1.3
omegaconf==2.3.0
hydra-core==1.3.2
gymnasium==0.27.1
pyquaternion==0.9.9
torch-gpytorch
numpy-quaternion